refactor(py/genkit/core/action): split up large module into smaller pieces and clean up internal namespace references
CHANGELOG:
- [ ] Fix all the references to internal namespaces in plugins and
samples.
- [ ] Export action types using `genkit.core.action.types`
- [ ] Split the `action.py` module into several files.
- [ ] Update all references to `genkit.core.action` and export
  only critical bits.
- [ ] Extract some action arg parsing into a utility function.
