+++ py/packages/genkit-ai/src/genkit/ai/veneer.py
@@ -115,7 +115,8 @@ from genkit.blocks.model import (
     ModelMiddleware,
 )
 from genkit.blocks.prompt import to_generate_action_options
-from genkit.core.action import ActionKind, ActionRunContext
+from genkit.core.action import ActionRunContext
+from genkit.core.action.types import ActionKind
 from genkit.core.environment import is_dev_environment
 from genkit.core.reflection import make_reflection_server
 from genkit.typing import (
