+++ py/packages/genkit-ai/src/genkit/core/registry.py
@@ -33,15 +33,11 @@ from typing import Any
 
 from genkit.core.action import (
     Action,
-    ActionKind,
     create_action_key,
     parse_action_key,
     parse_plugin_name_from_action_name,
 )
-
-type ActionName = str
-
-type ActionResolver = Callable[[ActionKind, str], None]
+from genkit.core.action.types import ActionKind, ActionName, ActionResolver
 
 # An action store is a nested dictionary mapping ActionKind to a dictionary of
 # action names and their corresponding Action instances.
@@ -122,7 +118,8 @@ class Registry:
             kind: The type of action being registered (e.g., TOOL, MODEL).
             name: A unique name for the action within its kind.
             fn: The function to be called when the action is executed.
-            metadata_fn: The function to be used to infer metadata (e.g. schemas).
+            metadata_fn: The function to be used to infer metadata (e.g.
+                schemas).
             description: Optional human-readable description of the action.
             metadata: Optional dictionary of metadata about the action.
             span_metadata: Optional dictionary of tracing span metadata.
