+++ py/plugins/compat-oai/src/genkit/plugins/compat_oai/models/handler.py
@@ -24,8 +24,7 @@ from typing import Any
 
 from openai import OpenAI
 
-from genkit.ai.registry import GenkitRegistry
-from genkit.core.action import ActionRunContext
+from genkit.ai import ActionRunContext, GenkitRegistry
 from genkit.plugins.compat_oai.models.model import OpenAIModel
 from genkit.plugins.compat_oai.models.model_info import (
     SUPPORTED_OPENAI_MODELS,
