+++ py/plugins/vertex-ai/src/genkit/plugins/vertex_ai/imagen.py
@@ -21,7 +21,7 @@ from typing import Any, Literal
 from pydantic import BaseModel
 from vertexai.preview.vision_models import ImageGenerationModel
 
-from genkit.core.action import ActionRunContext
+from genkit.ai import ActionRunContext
 from genkit.typing import (
     GenerateRequest,
     GenerateResponse,
