+++ py/plugins/vertex-ai/tests/test_gemini.py
@@ -18,7 +18,7 @@
 
 import pytest
 
-from genkit.core.action import ActionRunContext
+from genkit.ai import ActionRunContext
 from genkit.plugins.vertex_ai.gemini import Gemini, GeminiVersion
 from genkit.typing import (
     GenerateRequest,
